+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to the bonus-bot project will be documented in this file. The format is based
 on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.0.4] - 2023-09-16
+
+### Added
+
+- Added the ability to post in Chaohua.
+
+## [0.0.3] - 2023-03-18
+
+### Added
+
+- Added the ability to forward Weibo.
+
 ## [0.0.2] - 2023-02-25
 
 ### Changed